pyproject.toml
[project]
name = "pythongrpclearner"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "daphne>=4.2.1",
    "django>=5.2.6",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
]
